refactor: Split contact rework and filter in two seperate comp
